+++ src/php/Gdbots/Schemas/Analytics/Mixin/Tracker/Tracker.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Analytics\Mixin\Tracker;
+
+use Gdbots\Pbj\Message;
+
+interface Tracker extends Message
+{
+}