+++ src/php/Gdbots/Schemas/Analytics/Mixin/Tracker/TrackerV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Analytics\Mixin\Tracker;
+
+interface TrackerV1 extends Tracker
+{
+}