+++ src/php/Gdbots/Schemas/Analytics/Tracker/Keen.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Analytics\Tracker;
+
+use Gdbots\Pbj\Message;
+
+interface Keen extends Message
+{
+}